Decouple creating weight matrix
The weight matrix is now made separately from running the sim, so we
can run many sims on the same network with different stimuli. The
network parameters (weights, taus, gEL, Nns) get saved and loaded as
txt files. runsompsim reads the PBS ARRAY ID so each job will
correspond to a different stimulus. sompsim now returns stimrates which
is the associated stimulus.

(Theoretically. right now all the jobs in the array get killed for some
reason. it seems to be able to handle one successfully though. results
of that in times0.txt and spikes0.txt)

Nns is called such so it doesn’t get confused with ns (number of
spikes). It is the three N variables Ne, Ni, and N0.

deleting data from the github

see README in data folder on machine for new naming conventions